+++ entity-view/testsuite/src/test/java/com/blazebit/persistence/view/testsuite/limit/first/model/DocumentLimitView.java
@@ -1,0 +1,22 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ * Copyright Blazebit
+ */
+
+package com.blazebit.persistence.view.testsuite.limit.first.model;
+
+import com.blazebit.persistence.testsuite.entity.Document;
+import com.blazebit.persistence.view.EntityView;
+import com.blazebit.persistence.view.testsuite.basic.model.IdHolderView;
+
+/**
+ *
+ * @author Christian Beikov
+ * @since 1.5.0
+ */
+@EntityView(Document.class)
+public interface DocumentLimitView extends IdHolderView<Long> {
+
+    public String getName();
+
+}