+++ entity-view/testsuite/src/test/java/com/blazebit/persistence/view/testsuite/limit/first/model/PersonLimitJoinExpressionView.java
@@ -3,11 +3,10 @@
  * Copyright Blazebit
  */
 
-package com.blazebit.persistence.view.testsuite.limit.model;
+package com.blazebit.persistence.view.testsuite.limit.first.model;
 
 import com.blazebit.persistence.testsuite.entity.Person;
 import com.blazebit.persistence.view.EntityView;
-import com.blazebit.persistence.view.FetchStrategy;
 import com.blazebit.persistence.view.Limit;
 import com.blazebit.persistence.view.Mapping;
 
@@ -19,10 +18,10 @@ import java.util.List;
  * @since 1.5.0
  */
 @EntityView(Person.class)
-public interface PersonLimitMultisetView extends PersonLimitView {
+public interface PersonLimitJoinExpressionView extends PersonLimitView {
 
     @Limit(limit = "1", order = {"age", "id"})
-    @Mapping(fetch = FetchStrategy.MULTISET)
+    @Mapping("ownedDocuments[owner.name = VIEW(name)]")
     public List<DocumentLimitView> getOwnedDocuments();
 
 }