+++ entity-view/testsuite/src/test/java/com/blazebit/persistence/view/testsuite/limit/first/model/PersonLimitView.java
@@ -1,0 +1,23 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ * Copyright Blazebit
+ */
+
+package com.blazebit.persistence.view.testsuite.limit.first.model;
+
+import com.blazebit.persistence.view.testsuite.basic.model.IdHolderView;
+
+import java.util.List;
+
+/**
+ *
+ * @author Christian Beikov
+ * @since 1.5.0
+ */
+public interface PersonLimitView extends IdHolderView<Long> {
+
+    public String getName();
+
+    public List<DocumentLimitView> getOwnedDocuments();
+
+}