+++ entity-view/testsuite/src/test/java/com/blazebit/persistence/view/testsuite/limit/topn/model/PersonLimitJoinView.java
@@ -3,23 +3,23 @@
  * Copyright Blazebit
  */
 
-package com.blazebit.persistence.view.testsuite.limit.model;
+package com.blazebit.persistence.view.testsuite.limit.topn.model;
+
+import java.util.List;
 
 import com.blazebit.persistence.testsuite.entity.Person;
 import com.blazebit.persistence.view.EntityView;
 import com.blazebit.persistence.view.Limit;
 
-import java.util.List;
-
 /**
  *
  * @author Christian Beikov
- * @since 1.5.0
+ * @since 1.6.16
  */
 @EntityView(Person.class)
 public interface PersonLimitJoinView extends PersonLimitView {
 
-    @Limit(limit = "1", order = {"age", "id"})
+    @Limit(limit = "2", order = {"age", "id"})
     public List<DocumentLimitView> getOwnedDocuments();
 
 }