+++ entity-view/testsuite/src/test/java/com/blazebit/persistence/view/testsuite/limit/topn/model/PersonLimitSubselectView.java
@@ -3,7 +3,9 @@
  * Copyright Blazebit
  */
 
-package com.blazebit.persistence.view.testsuite.limit.model;
+package com.blazebit.persistence.view.testsuite.limit.topn.model;
+
+import java.util.List;
 
 import com.blazebit.persistence.testsuite.entity.Person;
 import com.blazebit.persistence.view.EntityView;
@@ -11,18 +13,16 @@ import com.blazebit.persistence.view.FetchStrategy;
 import com.blazebit.persistence.view.Limit;
 import com.blazebit.persistence.view.Mapping;
 
-import java.util.List;
-
 /**
  *
  * @author Christian Beikov
- * @since 1.5.0
+ * @since 1.6.16
  */
 @EntityView(Person.class)
-public interface PersonLimitSelectView extends PersonLimitView {
+public interface PersonLimitSubselectView extends PersonLimitView {
 
-    @Limit(limit = "1", order = {"age", "id"})
-    @Mapping(fetch = FetchStrategy.SELECT)
+    @Limit(limit = "2", order = {"age", "id"})
+    @Mapping(fetch = FetchStrategy.SUBSELECT)
     public List<DocumentLimitView> getOwnedDocuments();
 
 }